Update AbstractRepository's Interface
